split container into smaller modules
